pydrawbot/requirements.txt
svg.path
pyserial
